moved user disconnect handling to gateway
